chore: refactor loop details
