+++ src/apps/Loop/components/Sidesheet/LoopSidesheet.tsx
@@ -1,6 +1,5 @@
 import { Progress, Tabs } from '@equinor/eds-core-react';
 import { tokens } from '@equinor/eds-tokens';
-import { formatDateString, StringCell, Table } from '@equinor/GardenUtils';
 import { SidesheetApi } from '@equinor/sidesheet';
 import { useEffect, useState } from 'react';
 import { useQuery } from 'react-query';
@@ -11,6 +10,7 @@ import { generateExpressions, generateFamRequest } from '../../utility/helpers/f
 import { Banner } from './Banner';
 import { BannerItem } from './BannerItem';
 import { LoopContentTable } from './LoopContentTable';
+import { LoopDetails } from './LoopDetails';
 import { LoopWorkOrderTab } from './LoopWorkorderTable';
 
 type LoopSidesheetProps = {
@@ -89,76 +89,3 @@ export const LoopSidesheet = ({ item, actions }: LoopSidesheetProps) => {
 export const SidesheetTabList = styled(Tabs.List)`
     background-color: ${tokens.colors.ui.background__light.hex};
 `;
-
-type LoopDetailsProps = {
-    loop: Loop;
-};
-const LoopDetails = ({ loop }: LoopDetailsProps) => {
-    return (
-        <div>
-            <h3>Details</h3>
-            <Table>
-                <tbody>
-                    <tr>
-                        <td>Project</td>
-                        <td>
-                            <StringCell value={loop.project} />
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>Loop</td>
-                        <td>
-                            <StringCell value={loop.tagNo} />
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>Comm pkg</td>
-                        <td>
-                            <StringCell value={loop.commissioningPackageNo} />
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>Mc pkg</td>
-                        <td>
-                            <StringCell value={loop.mechanicalCompletionPackageNo} />
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>Area</td>
-                        <td>
-                            <StringCell value={loop.location} />
-                        </td>
-                    </tr>
-                    <tr>
-                        <td>Aggregated MC status</td>
-                        <td>
-                            <StringCell value={loop.loopContentStatus} />
-                        </td>
-                    </tr>
-
-                    <tr>
-                        <td>Planned/Actual MC complete</td>
-                        <td>
-                            <StringCell
-                                value={formatDateString(
-                                    loop.woActualCompletionDate
-                                        ? loop.woActualCompletionDate.toString()
-                                        : null
-                                )}
-                            />
-                        </td>
-                    </tr>
-
-                    <tr>
-                        <td>Remaning hours</td>
-                        <td>
-                            <StringCell
-                                value={loop.remainingManHours ? `${loop.remainingManHours}` : 'N/A'}
-                            />
-                        </td>
-                    </tr>
-                </tbody>
-            </Table>
-        </div>
-    );
-};
